--- instaclone/database/alembic/versions/41185fea0a70_.py
@@ -1,38 +1,0 @@
-"""empty message
-
-Revision ID: 41185fea0a70
-Revises: 86d24f3befb5
-Create Date: 2025-01-09 01:44:13.017647
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '41185fea0a70'
-down_revision: Union[str, None] = '86d24f3befb5'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('blocked_token',
-    sa.Column('id', sa.BigInteger(), autoincrement=True, nullable=False),
-    sa.Column('jti', sa.String(length=40), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.drop_constraint('media_ibfk_2', 'media', type_='foreignkey')
-    op.create_foreign_key(None, 'media', 'stories', ['story_id'], ['story_id'], ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'media', type_='foreignkey')
-    op.create_foreign_key('media_ibfk_2', 'media', 'stories', ['story_id'], ['story_id'])
-    op.drop_table('blocked_token')
-    # ### end Alembic commands ###
