--- instaclone/database/alembic/versions/86d24f3befb5_.py
@@ -1,52 +1,0 @@
-"""empty message
-
-Revision ID: 86d24f3befb5
-Revises: 18e806f94e68
-Create Date: 2025-01-04 11:09:25.551137
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = '86d24f3befb5'
-down_revision: Union[str, None] = '18e806f94e68'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('users', 'gender',
-               existing_type=mysql.VARCHAR(length=10),
-               nullable=True)
-    op.alter_column('users', 'birthday',
-               existing_type=sa.DATE(),
-               nullable=True)
-    op.alter_column('users', 'introduce',
-               existing_type=mysql.VARCHAR(length=100),
-               nullable=True)
-    op.alter_column('users', 'website',
-               existing_type=mysql.VARCHAR(length=100),
-               nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('users', 'website',
-               existing_type=mysql.VARCHAR(length=100),
-               nullable=False)
-    op.alter_column('users', 'introduce',
-               existing_type=mysql.VARCHAR(length=100),
-               nullable=False)
-    op.alter_column('users', 'birthday',
-               existing_type=sa.DATE(),
-               nullable=False)
-    op.alter_column('users', 'gender',
-               existing_type=mysql.VARCHAR(length=10),
-               nullable=False)
-    # ### end Alembic commands ###
