+++ instaclone/database/alembic/versions/d9846c7e8030_db_initialized.py
@@ -1,0 +1,144 @@
+"""DB initialized
+
+Revision ID: d9846c7e8030
+Revises: 
+Create Date: 2025-01-17 13:23:06.067010
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'd9846c7e8030'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('blocked_token',
+    sa.Column('id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('jti', sa.String(length=40), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('users',
+    sa.Column('user_id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('username', sa.String(length=20), nullable=False),
+    sa.Column('password', sa.String(length=128), nullable=False),
+    sa.Column('full_name', sa.String(length=30), nullable=False),
+    sa.Column('email', sa.String(length=100), nullable=False),
+    sa.Column('phone_number', sa.String(length=11), nullable=False),
+    sa.Column('creation_date', sa.Date(), nullable=False),
+    sa.Column('profile_image', sa.String(length=100), nullable=False),
+    sa.Column('gender', sa.String(length=10), nullable=True),
+    sa.Column('birthday', sa.Date(), nullable=True),
+    sa.Column('introduce', sa.String(length=100), nullable=True),
+    sa.Column('website', sa.String(length=100), nullable=True),
+    sa.PrimaryKeyConstraint('user_id'),
+    sa.UniqueConstraint('email'),
+    sa.UniqueConstraint('phone_number'),
+    sa.UniqueConstraint('username')
+    )
+    op.create_table('followers',
+    sa.Column('follower_id', sa.BigInteger(), nullable=False),
+    sa.Column('following_id', sa.BigInteger(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['follower_id'], ['users.user_id'], ),
+    sa.ForeignKeyConstraint(['following_id'], ['users.user_id'], ),
+    sa.PrimaryKeyConstraint('follower_id', 'following_id')
+    )
+    op.create_table('posts',
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('post_id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('location', sa.String(length=50), nullable=True),
+    sa.Column('post_text', sa.String(length=500), nullable=True),
+    sa.Column('creation_date', sa.DATETIME(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.user_id'], ),
+    sa.PrimaryKeyConstraint('post_id')
+    )
+    op.create_table('stories',
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('story_id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('creation_date', sa.DATETIME(), nullable=False),
+    sa.Column('expiration_date', sa.DATETIME(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.user_id'], ),
+    sa.PrimaryKeyConstraint('story_id')
+    )
+    op.create_table('comments',
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('post_id', sa.BigInteger(), nullable=False),
+    sa.Column('comment_id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('parent_id', sa.BigInteger(), nullable=True),
+    sa.Column('comment_text', sa.String(length=200), nullable=False),
+    sa.ForeignKeyConstraint(['parent_id'], ['comments.comment_id'], ),
+    sa.ForeignKeyConstraint(['post_id'], ['posts.post_id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.user_id'], ),
+    sa.PrimaryKeyConstraint('comment_id')
+    )
+    op.create_table('media',
+    sa.Column('post_id', sa.BigInteger(), nullable=True),
+    sa.Column('story_id', sa.BigInteger(), nullable=True),
+    sa.Column('image_id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('file_name', sa.String(length=100), nullable=False),
+    sa.Column('url', sa.String(length=200), nullable=False),
+    sa.ForeignKeyConstraint(['post_id'], ['posts.post_id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['story_id'], ['stories.story_id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('image_id')
+    )
+    op.create_table('post_likes',
+    sa.Column('like_id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('content_id', sa.BigInteger(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['content_id'], ['posts.post_id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.user_id'], ),
+    sa.PrimaryKeyConstraint('like_id')
+    )
+    op.create_table('story_likes',
+    sa.Column('like_id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('content_id', sa.BigInteger(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['content_id'], ['stories.story_id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.user_id'], ),
+    sa.PrimaryKeyConstraint('like_id')
+    )
+    op.create_table('story_views',
+    sa.Column('story_view_id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('story_id', sa.BigInteger(), nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('viewed_at', sa.DATETIME(), nullable=False),
+    sa.ForeignKeyConstraint(['story_id'], ['stories.story_id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.user_id'], ),
+    sa.PrimaryKeyConstraint('story_view_id')
+    )
+    op.create_table('comment_likes',
+    sa.Column('like_id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('content_id', sa.BigInteger(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['content_id'], ['comments.comment_id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.user_id'], ),
+    sa.PrimaryKeyConstraint('like_id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('comment_likes')
+    op.drop_table('story_views')
+    op.drop_table('story_likes')
+    op.drop_table('post_likes')
+    op.drop_table('media')
+    op.drop_table('comments')
+    op.drop_table('stories')
+    op.drop_table('posts')
+    op.drop_table('followers')
+    op.drop_table('users')
+    op.drop_table('blocked_token')
+    # ### end Alembic commands ###
